uncommited work on the refactor
